Wifi: Clear connected device list cache when SoftAP is turned  OFF.
When Tethering module receives WIFI_AP_STATE_CHANGED indication it clears
cached entries for connected device list. This part of the logic fails to
execute as this indication is also monitored with the latest upgrade.
Address this issue by ensuring the clear to happen on the
WI_FI_AP_STATE_CHANGED indication.

CRs-Fixed: 2034431
Change-Id: I15cc32478992bc305f563a7fbbf55f1e43a030fe